moved around js source files
